Split app/import.go in multiple files
